+++ src/components/common/Navbar/Navbar.tsx
@@ -7,18 +7,13 @@ import CartToggle from 'src/components/cart/CartToggle'
 import IconButton from 'src/components/ui/IconButton'
 import Link from 'src/components/ui/Link'
 import Logo from 'src/components/ui/Logo'
-import {
-  CaretLeft as CaretLeftIcon,
-  List as ListIcon,
-  X as XIcon,
-} from 'phosphor-react'
 import SignInLink from 'src/components/ui/SignInLink'
 import SlideOver from 'src/components/ui/SlideOver'
 import { useStoreCollection } from 'src/hooks/useAllCollections'
 import { mark } from 'src/sdk/tests/mark'
 import PostalCodeInput from 'src/components/common/PostalCode'
-
-import SearchInput from '../SearchInput'
+import IconSVG from 'src/components/common/IconSVG'
+import SearchInput from 'src/components/common/SearchInput'
 
 import './navbar.scss'
 
@@ -67,7 +62,7 @@ function Navbar() {
               <IconButton
                 classes="navbar__menu"
                 aria-label="Open Menu"
-                icon={<ListIcon size={32} />}
+                icon={<IconSVG name="List" width={32} height={32} />}
                 onClick={() => setShowMenu(true)}
               />
               <LinkGatsby
@@ -89,7 +84,7 @@ function Navbar() {
               <IconButton
                 classes="navbar__collapse"
                 aria-label="Collapse search bar"
-                icon={<CaretLeftIcon size={32} />}
+                icon={<IconSVG name="CaretLeft" width={32} height={32} />}
                 onClick={() => setSearchExpanded(false)}
               />
             )}
@@ -133,7 +128,7 @@ function Navbar() {
             <IconButton
               classes="navbar__button"
               aria-label="Close Menu"
-              icon={<XIcon size={32} />}
+              icon={<IconSVG name="X" width={32} height={32} />}
               onClick={() => dismissTransition.current?.()}
             />
           </header>